refactor: extract flag type definitions into deno_cli_parser crate
Introduce a new `libs/cli_parser` (`deno_cli_parser`) crate that holds
the canonical CLI flag type definitions (`Flags`, `DenoSubcommand`,
`PermissionFlags`, all the `*Flags` structs/enums, and their pure
data impls). `cli/args/flags.rs` now re-exports these via
`pub use deno_cli_parser::flags::*` and keeps all clap-based parsing
in place unchanged.

This is the first step of splitting the larger effort to replace clap
with a zero-cost parser. It is a pure refactor with no behavior change:
clap still does all argument parsing; only the type definitions moved.

CLI-specific helper methods that depend on clap, deno_graph,
deno_telemetry, or the cli build environment (env!("TARGET")) cannot
live in the parser crate, so they become extension traits in
`cli/args/flags.rs`: FileFlagsExt, CompileFlagsExt, DenoSubcommandExt,
TypeCheckModeExt, and FlagsExt. Consumers gained the corresponding
`use crate::args::*Ext` imports.

`HelpFlags.help` changes from `clap::builder::StyledStr` to `String`
(the ansi-rendered string is now stored at construction) so the parser
crate does not depend on clap.

The `From<BundleOptions> for BundleFlags` conversion moves into the
crate alongside `BundleFlags` to satisfy the orphan rule.
