+++ cli/factory.rs
@@ -66,7 +66,9 @@ use crate::args::CliLockfile;
 use crate::args::CliOptions;
 use crate::args::ConfigFlag;
 use crate::args::DenoSubcommand;
+use crate::args::DenoSubcommandExt;
 use crate::args::Flags;
+use crate::args::FlagsExt;
 use crate::args::InstallFlags;
 use crate::args::InstallFlagsLocal;
 use crate::cache::Caches;
