+++ cli/tools/bench/mod.rs
@@ -37,6 +37,7 @@ use tokio::sync::mpsc::unbounded_channel;
 use crate::args::BenchFlags;
 use crate::args::CliOptions;
 use crate::args::Flags;
+use crate::args::TypeCheckModeExt;
 use crate::colors;
 use crate::display::write_json_to_stdout;
 use crate::factory::CliFactory;
