+++ cli/tools/x.rs
@@ -18,6 +18,7 @@ use node_resolver::BinValue;
 
 use crate::args::DenoXShimName;
 use crate::args::Flags;
+use crate::args::FlagsExt;
 use crate::args::XFlags;
 use crate::args::XFlagsKind;
 use crate::factory::CliFactory;
