+++ libs/cli_parser/Cargo.toml
@@ -1,0 +1,24 @@
+# Copyright 2018-2026 the Deno authors. MIT license.
+
+[package]
+name = "deno_cli_parser"
+version = "0.1.0"
+authors.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+description = "Zero-cost CLI argument parser for the Deno CLI"
+
+[lib]
+path = "src/lib.rs"
+
+[dependencies]
+deno_bundle_runtime.workspace = true
+deno_config = { workspace = true, features = ["workspace"] }
+deno_core.workspace = true
+deno_lib.workspace = true
+deno_npm_installer.workspace = true
+deno_runtime.workspace = true
+deno_semver.workspace = true
+log.workspace = true
+serde = { workspace = true, features = ["derive"] }
